change yaku-util.js to yaku-util.ts
